+++ typings/index.d.ts
@@ -1,4 +1,5 @@
 /// <reference path="./app.d.ts" />
 /// <reference path="./page.d.ts" />
+/// <reference path="./component.d.ts" />
 /// <reference path="./apis.d.ts" />
 /// <reference path="./event.d.ts" />